LayoutTests/imported/w3c/web-platform-tests/shadow-dom/declarative/tentative/shadowrootadoptedstylesheets/shadowrootadoptedstylesheets-async-fetch-mutation-reinsert-same-expected.txt


FAIL Placeholder removed and reinserted at same position is replaced by fetched sheet. assert_equals: Before removal: expected 1 placeholder(s). expected 1 but got 0
